fix(tests): três defeitos de SQL que abortavam suíte inteira (SCRUM-361)
Com o contexto de backend no lugar, as suítes passaram a rodar mais fundo e
mostraram erro próprio — cada um abortando o arquivo, não falhando uma asserção.

1. `get_sales_metrics`: `is(numeric, integer)`. A soma de literais
   `1000 + 500 + ...` é INTEGER, e `is()` do pgTAP é `anyelement, anyelement` —
   o Postgres nem acha a função. Cast explícito.

2. `get_funnel_flow`: "aggregate function calls cannot contain window function
   calls". `bool_and(rc >= lead(rc) OVER (...))` precisa de um nível próprio
   para a janela.

3. `get_commission_ledger`: "permission denied to set parameter
   session_replication_role". É parâmetro de superusuário, e a suíte roda como
   `service_role` a partir do contexto de backend. Volta a `postgres` só para
   desarmar os gatilhos e devolve o papel em seguida.

+++ supabase/tests/get_sales_metrics_test.sql
@@ -228,7 +228,11 @@ SELECT is(
 SELECT is(
   (public.get_sales_metrics('99599599-aaaa-0000-0000-000000000995','month','2027-07-15',
       NULL, NULL, NULL, '99599599-aaaa-2222-0000-000000000995') ->> 'revenue_total')::numeric,
-  1000 + 500 + 7777 + 111 + 400,  -- s1,s2,s6,s7,s8 do closer1
+  -- `::numeric` explícito: `is()` do pgTAP é `anyelement, anyelement`, e a soma
+  -- de literais inteiros dá INTEGER. Sem o cast, o par vira
+  -- `is(numeric, integer)` e o Postgres nem acha a função — a suíte aborta ali,
+  -- sem chegar às nove asserções seguintes.
+  (1000 + 500 + 7777 + 111 + 400)::numeric,  -- s1,s2,s6,s7,s8 do closer1
   '(g) filtro por closer1 soma só as vendas dele (sale_responsible_id)');
 
 -- ---------------------------------------------------------------------------
